requirements_pipreqs.txt
albumentations==2.0.5
matplotlib==3.10.1
moviepy==2.1.2
opencv_python==4.10.0.84
Pillow==11.1.0
playsound==1.3.0
pyfiglet==1.0.2
pygame==2.6.1
sahi==0.11.22
schedule==1.2.2
termcolor==3.0.0
twilio==9.5.1
ultralytics==8.3.97
